Updated some unit tests

+++ src/main/java/photoman/service/impl/ProfileServiceImpl.java
@@ -28,13 +28,13 @@ public class ProfileServiceImpl implements ProfileService
 	
 	@Transactional
 	@Override
-	public Profile createProfile(String profileName) throws ProfileException
+	public Profile createProfile(String profileName)
 	{
 		if(profileRepo.findByProfileName(profileName) != null)
 		{
 			throw ProfileException.getProfileExistsException(profileName);
 		}
-		setCurrentProfile(new Profile(profileName));
+		setCurrentProfile(profileRepo.save(new Profile(profileName)));
 		return currentProfile;
 	}
 	
@@ -51,6 +51,7 @@ public class ProfileServiceImpl implements ProfileService
 			{
 				//If we've never set a profile, then use the default one
 				profileProp = new Property(CURRENT_PROFILE_PROPERTY, DEFAULT_PROFILE_NAME);
+				propertyRepo.save(profileProp);
 			}
 			
 			//Get the profile based on the current profile property value
@@ -59,7 +60,7 @@ public class ProfileServiceImpl implements ProfileService
 			if(profile == null)
 			{
 				//If the default profile doesn't exist yet, then create it
-				profile = new Profile(profileName);
+				profile = profileRepo.save(new Profile(profileName));
 			}
 			setCurrentProfile(profile);
 		}
@@ -67,7 +68,8 @@ public class ProfileServiceImpl implements ProfileService
 	}
 	
 	@Transactional
-	public Profile changeProfile(String profileName) throws ProfileException
+	@Override
+	public Profile changeProfile(String profileName)
 	{
 		Profile profile = profileRepo.findByProfileName(profileName);
 		if(profile == null)
@@ -79,19 +81,36 @@ public class ProfileServiceImpl implements ProfileService
 	}
 	
 	@Transactional
-	public void deleteProfile(String profileName) throws ProfileException
+	@Override
+	public void deleteProfile(String profileName)
 	{
+		//First check if the profile we're deleting exists 
 		Profile profile = profileRepo.findByProfileName(profileName);
 		if(profile == null)
 		{
 			throw ProfileException.getProfileDoesntExistException(profileName);
 		}
+		
+		//Then check to make sure we're not deleting the last profile
+		if(profileRepo.count() <= 1)
+		{
+			throw ProfileException.getDeleteLastProfileException(profileName);
+		}
+
+		//Change to the default profile if the profile being deleted 
+		//is the current one - recreate the default profile if required
+		this.changeProfile(profileName);
+		//Delete the profile
 		profileRepo.delete(profile);
+		
+
 	}
 	
 	private void setCurrentProfile(Profile profile)
 	{
 		this.currentProfile = profile;
-		propertyRepo.save(new Property(CURRENT_PROFILE_PROPERTY, profile.getProfileName()));
+		Property profileProp = propertyRepo.findByPropName(CURRENT_PROFILE_PROPERTY);
+		profileProp.setPropValue(profile.getProfileName());
+		propertyRepo.save(profileProp);
 	}
 }
